fix: notification bug and limit quantity of them
